requirements-devel.txt
-r requirements.txt

ipython
ipdb

django-sampledatahelper==0.3
